+++ src/init.c
@@ -143,6 +143,9 @@ extern r_obj* ffi_locate_matches(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*,
 extern r_obj* ffi_locate_interval_merge_bounds(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_locate_interval_merge_groups(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_interval_complement(r_obj*, r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_check_list(r_obj*, r_obj*);
+extern r_obj* ffi_list_all_vectors(r_obj*);
+extern r_obj* ffi_list_check_all_vectors(r_obj*, r_obj*);
 
 
 // Maturing
@@ -307,6 +310,9 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_locate_interval_merge_bounds",      (DL_FUNC) &ffi_locate_interval_merge_bounds, 4},
   {"ffi_locate_interval_merge_groups",      (DL_FUNC) &ffi_locate_interval_merge_groups, 4},
   {"ffi_interval_complement",               (DL_FUNC) &ffi_interval_complement, 4},
+  {"ffi_check_list",                        (DL_FUNC) &ffi_check_list, 2},
+  {"ffi_list_all_vectors",                  (DL_FUNC) &ffi_list_all_vectors, 2},
+  {"ffi_list_check_all_vectors",            (DL_FUNC) &ffi_list_check_all_vectors, 2},
   {NULL, NULL, 0}
 };